refactor: Remove `singer_sdk._singerlib`
