+++ tests/singerlib/encoding/test_msgspec.py
@@ -9,14 +9,14 @@ from textwrap import dedent
 
 import pytest
 
-from singer_sdk._singerlib import RecordMessage
-from singer_sdk._singerlib.exceptions import InvalidInputLine
 from singer_sdk.contrib.msgspec import (
     MsgSpecReader,
     MsgSpecWriter,
     dec_hook,
     enc_hook,
 )
+from singer_sdk.singerlib import RecordMessage
+from singer_sdk.singerlib.exceptions import InvalidInputLine
 
 CALLBACKS = {
     "RECORD": lambda _: None,
