+++ tests/singerlib/test_messages.py
@@ -7,7 +7,7 @@ from contextlib import redirect_stdout
 import pytest
 from pytz import timezone
 
-import singer_sdk._singerlib as singer
+import singer_sdk.singerlib as singer
 
 UTC = datetime.timezone.utc
