+++ tests/singerlib/test_utils.py
@@ -5,8 +5,8 @@ from datetime import datetime, timezone
 import pytest
 import pytz
 
-from singer_sdk._singerlib import strftime, strptime_to_utc
-from singer_sdk._singerlib.utils import NonUTCDatetimeError
+from singer_sdk.singerlib import strftime, strptime_to_utc
+from singer_sdk.singerlib.utils import NonUTCDatetimeError
 
 
 def test_small_years():
